refactor: pydantic 문법 v1 -> v2로 변경 및 import 경로 수정 놓친거 수정
